validacion de parametros nulos

+++ app/Http/Controllers/SurveyController.php
@@ -8,6 +8,9 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
 use App\Models\Survey;
+use App\Models\Question;
+use App\Models\DefinedAnswer;
+use App\Models\NextQuestion;
 use App\Http\Controllers\TeamController;
 
 class SurveyController extends Controller
@@ -45,12 +48,12 @@ class SurveyController extends Controller
         $user = auth()->user();
 
         $validator = Validator::make($request->all(), [
-            'team_id' => 'integer',
+            'team_id' => 'integer|nullable',
             'privacy_id' => 'required|integer',
             'survey_name' => 'required|string|max:220',
-            'survey_description' => 'string',
-            'button_color' => 'string|max:7',
-            'background_color' => 'string|max:7',
+            'survey_description' => 'string|nullable',
+            'button_color' => 'string|max:7|nullable',
+            'background_color' => 'string|max:7|nullable',
         ]);
         if($validator->fails())
             return response()->json($validator->errors()->toJson(),400);
@@ -82,9 +85,9 @@ class SurveyController extends Controller
         $validator = Validator::make($request->all(), [
             'privacy_id' => 'required|integer',
             'survey_name' => 'required|string|max:220',
-            'survey_description' => 'string',
-            'button_color' => 'string|max:7',
-            'background_color' => 'string|max:7',
+            'survey_description' => 'string|nullable',
+            'button_color' => 'string|max:7|nullable',
+            'background_color' => 'string|max:7|nullable',
             'active'=>'required|integer',
         ]);
 
@@ -108,14 +111,30 @@ class SurveyController extends Controller
         $user = auth()->user();
         $team_ids = getTeamIds();
         $t = Survey::where('survey_id',$survey_id)
-            ->whereIn('team_id', $team_ids)
-            ->orWhere('user_id',$user->user_id)
+            ->where('user_id',$user->user_id)
             ->first();
+        if(!empty($t)){
+            $quest = Question::where('survey_id',$survey_id)
+                    ->select('question_id')
+                    ->get();
+            $qArr = [];
+            foreach ($quest as $q) {
+              array_push($qArr, $q->question_id); 
+            }
 
-        if($t->delete()>0) {
-            return response()->json([
-                'message' =>'Team successfully deleted!'
-            ],201);
+            $defAnsw = DefinedAnswer::whereIn('question_id',$qArr)
+                    ->delete();
+            $nextQuest = NextQuestion::whereIn('question_id',$qArr)
+                    ->delete();
+
+            $q = Question::where('survey_id',$survey_id)
+                    ->select('question_id')
+                    ->delete();
+            if($t->delete()>0) {
+                return response()->json([
+                    'message' =>'Survey successfully deleted!'
+                ],201);
+            }
         }
 
         return response()->json([
